correción de errores de tipado en ProductDetailPanel y useFieldValidation, además de agregar la carga de variantes en el panel de detalles del producto.
